Move runtime variables into their respective binary's symbol tables
It's still useful to know what's at a particular address when reverse
engineering a binary, even if it's a dynamically populated construct, so
it makes more sense for these to be together with the binary's symbol
tables if applicable, rather than in ram.yml. Now, reserve ram.yml for
addresses that don't fall in the address ranges of any binary.