Add function sync check between symbols and headers
Add a script to catch when the C headers define functions that don't
exist in the symbol tables. This might be the result of typos or from
renaming symbols. The script can be invoked with the symbol-check make
target from the `headers/` directory, and will be run automatically in
GitHub Actions.

Also fix the "TaskProcBoot" function signature that was caught by the
sync check script.