Add symbols_vfill.py and supporting tools
This is a command line utility that uses arm5find.py and offsets.py to
automatically fill in addresses in the pmdsky-debug symbol tables.

Also included in this commit is resymgen.py, which is a programmatic
Python interface for running resymgen commands. Additionally, fix the
ordering of binary options in the offsets.py help text.