Add JSON output format to resymgen (#24)
The JSON output format is similar to the ghidra output format, but more
complete (it includes length and description).

Also add a custom Ghidra script for importing from this JSON format. It
does the same thing as the built-in ImportSymbolsScript.py, except it
also adds plate comments using symbol descriptions.