Support 64-bit addresses and lengths in resymgen
There's no reason why resymgen shouldn't support 64-bit binaries, and as
a somewhat general-purpose tool it's arguably expected that it should.
Change Uint to u64 to allow this, and add a few tests to make sure
things still work with numbers larger-than-32-bit numbers.

Users working with 32-bit binaries should not be meaningfully affected
by this increase in maximum value, since specifying both the address and
length for each symbol block is already required, and should constrain
symbol addresses and lengths to 32-bit integers via the in-bounds
symbols check resymgen provides, even when the underlying Uint type is
u64.