Relicense symbols and C headers under GPLv3-MIT dual license
Projects like tech-ticks' c-of-time use the symbol tables and C headers
to generate actual ROM patches. However, the current GPLv3 licensing of
the symbols/headers are problematic for the distribution of these
patches, since they are inherently tied to the closed-source EoS ROMs
and thus incompatible with GPL.

Release the symbols and C headers under a dual GPLv3-MIT license to make
it easier for people to create and distribute patches derived from them.
Patch creators may choose to use the symbols and headers under the MIT
license.

Note that this commit does NOT change the license of the overall
pmdsky-debug repository, which is and remains GPLv3.