Make VersionDep value lookups more robust (#39)
VersionDeps map Versions to values, where a Version is an ordinal-string
pair. Within a Block, the ordinal for Version keys is determined by the
version list in the Block's metadata and makes it possible to sort
version keys according to the Block's global version list.

The problem is that intuitively, a version will often be conceptualized
as a vanilla string, without the ordinal (which is largely an
implementation detail). This leads to unexpected results when attempting
a lookup in a VersionDep or MaybeVersionDep with a Version key that
matches one of the native keys by string value, but not by ordinal (with
the end result being an unexpected return value of None). This is
especially error-prone when working with glue code between a SymGen and
foreign symbols (like with the `merge` command), which caused a bug in
bounds checking for versions with ordinals greater than 0 (i.e., not the
first version in the list).

Rename the old VersionDep/MaybeVersionDep key-lookup functions to
include a "_native" suffix that highlights the expectation that keys are
in the native key space, with proper names AND ordinals.

Reimplement the vanilla key-lookup functions (without "_native") to
actually iterate over a VersionDep/MaybeVersionDep's keys to find a
native key that matches by name, but not necessarily by ordinal. This
turns lookups into an O(n) operation as opposed to O(1) (which is why
the "_native" functions are still useful), but is much more intuitive
for general-purpose key lookups, where the argument may not necessarily
come from the native key space. This fixes the aforementioned
bounds-checking bug.