Use reverse chronological order for symbol versions
Explorers of Sky released, from latest to earliest, for EU, NA, and JP.
It makes sense to use the same order for symbol versions so that "newer"
versions take precedence for sorting, especially since it's likely the
EU binary added more new symbols than it removed.

The only symbols reordered relative to other symbols are the
EU-exclusive ones.