Add cursors for recursive subregion traversal
Traversing the nested hierarchy of subregions is quite unwieldy since
the subregion list on each block is optional, not all subregions are
guaranteed to be resolved, and file paths aren't stored directly on a
SymGen. Add a cursor API that makes such traversal more convenient.
Cursors will automatically keep track of details like file paths, and
whether or not subregions are resolved. When manipulating actual files
in a file system, this is what is needed in practice most of the time.