Dungeon struct research (#74)
Research and document all the big fields in the dungeon struct, which brings
down the length of dungeon.h down to ~2100 lines.

Also remove the OMIT_LARGE_FILES compiler define. This only existed
because dungeon.h was enormous and took forever to import into Ghidra;
now it isn't even the largest header file (some of the enum files are larger),
so the flag is no longer needed.

Also add a couple of return values that were missing from overlay29.yml.