Add generic ARM9/ARM7 information
We're bound to need an ARM7 file eventually if/when sound engine
work is documented. For now, just add the binary info and the
entrypoint. Also update offsets.py with the ARM7 binary.

While we're at it, add the ARM9 entrypoint, and add to the description
describing the general responsibilities of the ARM9 CPU, to match the
blurb in the ARM7 description.