Port overlay_0011 from Irdkwia's ASM.ods
Plus verification of a few of the symbols by me.

Note: The data symbol now labeled as OVERLAY11_UNKNOWN_TABLE__NA_2316A38
was listed in Irdkwia's notes as having 20 (typo'd as 2) 8-byte entries,
but from investigation, it seems like they're actually 4-byte entries.
The type is now listed as undefined4[40].