resymgen: Improve gen subcommand usability with unversioned addresses
Previously, to run the gen command with unversioned addresses, a user
would need to specify something like `-v ""` to force a dummy version
for the output files. While this works, it's not very intuitive.

Now, add version inference. If a symgen YAML file is completely
versionless, just generate one set of output files without version
suffixes by default (if no version names are explicitly specified).
Also add version inference for the case where a version list is not
specified, but addresses are versioned, although this may not work
properly in all cases (a warning will be displayed in these cases).