resymgen: Improve symbol list sorting with missing versions
Previously, if some symbols were missing addresses for certain versions,
they might artificially end up at the end when the formatter was run
(which sorts symbols by address). For example, given the following list:

- name: symbol2
  address:
    v1: 0
    v2: 1
- name: symbol1
  address:
    v2: 0

It seems intuitively obvious that symbol1 should come before symbol2
based on the v2 addresses, but the formatter would place symbol2 before
symbol1 because symbol1 lacks a v1 address.

Now, improve the sorting algorithm such that in cases like the above,
the sort order will be more intuitive (so symbol1 would be placed before
symbol2 in the above example). For more details on the actual sorting
algorithm, see the comments in SymbolList::sort().

As a bonus, the new algorithm also fixes the issue with intransitive
comparisons when a list contains both versioned and unversioned
addresses.