Info dump 24-26/66: PP, IQ and recruiting (#110)
Adds functions mostly related to:
- Using and recovering PP
- IQ skills
- Enemy recruiting in dungeons

Other relevant changes include:

The faint_reason enum has been renamed to damage_source, since I discovered that values of this enum are passed to CalcDamageFixed. This function checks if the damage source is equal to DAMAGE_SOURCE_THROWN_ITEM and increases the damage if the user has the Power Pitcher IQ skill, so it's more than just a faint reason (although many of the values are used as faint reasons)

The old damage_source enum has been renamed to damage_message, since it looks like it's only used to print damage-related messages to the message log.