Add support for unsized header inclusion by external apps
The redefinitions of various system types in pmdsky.h make it
self-sufficient, but can be problematic for external applications trying
to include it. For example, C++ doesn't allow you to typedef the bool
type.

This commit adds an alternate inclusion mode for pmdsky.h, activated
with the PMDSKY_UNSIZED_HEADERS preprocessor define. In this mode,
exact struct sizing is disregarded, and many types are pulled from the
standard C system headers. The intent is so that an external application
can do something like this:
```
 #define PMDSKY_UNSIZED_HEADERS
 #include "pmdsky-debug/headers/pmdsky.h"
 #undef PMDSKY_UNSIZED_HEADERS
```

The typedef of va_list is also moved into the pmdsky.h preamble, where
it should've been in the first place, since it's normally defined in
stdarg.h, which is a standard C header.

This change is backwards compatible. If PMDSKY_UNSIZED_HEADERS is not
defined, pmdsky.h will behave as before.