Add order checking to symbol_check.py
It's currently pretty easy for the function/data symbols in the C
headers to end up out of order relative to the actual symbol tables.
Enhance symbol_check.py to check for this and report ordering changes
that should be done.

Also make the necessary corrections for this new check to pass.