Defined unknown functions in ApplyItemEffect (#134)
Mostly changes related to figuring out what the functions in ApplyItemEffect did.

Also shrunk the dungeon::secret_item_room_weights and gave some of its space
to the trap_weights because the function 0x022E7A30 (NA) specifically starts
looking at DUNGEON_PTR + 0x02C932. The bounds on the check is 24 and that
happens to line up with the number of traps that can be placed during dungeon
generation (GRUDGE_TRAP).