Add arm5correlate.py utility
This script was first written to look for correlations between the DS
Download Play binary and the release binaries (#146). It has since
proven useful for correlating code across other binaries, so add it to
the repo's collection of tools.

Note that this script is quite complex and hasn't been carefully
quality-checked, so it might not be totally reliable.