Add headers-with-docstrings release artifact
It may be preferable to download the headers-with-docstrings artifact
instead of the normal headers artifact. Since add_header_docstrings.py
is somewhat hacky and experimental, and because the docstrings add a
bunch of potentially unneeded cruft to the headers, make a separate
parallel artifact rather than modifying the existing release artifact.