Some 3d engine documentation (#168)
The gist of this is the game uses the 3D rendering system to render
a few things. From what I’ve seen, it mostly includes weather and
text box border/background, but could be easily extended (and is
what Irdkwia used for the moving spirit effect in the "angry" part of
Tangent’s hack (that I can’t remember the name of).