Some arm7 interrupt documentation (#201)
I was working on some arm7 reversing and found some mirrored code
between arm9 and arm7 where EnableIrqFlag and SetIrqFlag were
being used, and from there I was able to find a few more arm9 interrupt
functions in the arm7 as well. Interestingly although all of these ones
are identical, the last interrupt function GetIrqFlag is not, and it does
an and with a different mask, so I didn't include it here.

Also a note for all future arm7 documentation: I did a checksum on the
arm7.bin of all three editions of EoS and they are all identical (SHA256 CB2A9B26D77A6E7D0ADD85C27E3642105918695D040D02F70D5EBC1D52C9495B),
and since they are also all loaded at 0x2380000, addresses for all
editions will be identical if the function is in the binary. There do
appear to be some functions that seem to be written in during runtime
though, so this might not apply for those ones.