Filled in missing EU addresses (#212)
With EU support in the decomp, it is now possible to precisely
identify the address of any NA/EU symbol when given an address
in the other game version. I updated the syncing script within the
decomp project to fill in any missing addresses between EU/NA
(mostly EU). There were a handful of cases where EU and NA were
pointing to different symbols, which I also ironed out.