Identify more system utilities
These were found by correlating identical system code between EoS and
Pokemon Heart Gold, for which various bits of library code (e.g., the
Nitro SDK) are already known.

This is the first of a series of results from this correlation work.
Some of the changes included in this commit:
- init code
- math utilities for fixed-point numbers, vectors, and matrices
- graphics utilities

Note: GX_HBlankIntr is not the same between EoS and Heart Gold, but the
label was added anyway because it is between GX_Init and GX_VBlankIntr
in both games. The latter two functions DO match between games.