Support symbol aliases in Ghidra import scripts
Create additional non-primary labels for symbol aliases at the same
address as the primary symbol names. Functions and data are treated the
same by the script; Ghidra's createLabel() function does the right
thing (i.e., if a function exists at that address, the label will be
marked as a function, otherwise it will be a data label).