Support the PYTHON_TERM_COLOR environment variable in tools
PYTHON_TERM_COLOR mimics the CARGO_TERM_COLOR environment variable,
which can be set to "always" to force colors on and "never" to force
colors off (by default, it will be treated as "auto" to auto-detect the
environment).

This will affect any Python tools that use the ansi.py module.