Touchscreen struct and script parameter conversion functions (#254)
This commit contains two minor additions.

First, I've documented a struct that can be used to read data
about the current status of the touchscreen, including whether
it's currently being pressed or not and the position being pressed.
This data, unlike the raw values that can be read from the I/O
ports, is expressed in pixels starting from the top left corner of
the screen, which should be a better spot to read touchscreen
coordinates since it's easier to use and consistent across
emulators (seems like the raw values are not).

Second, I've researched the function previously known as
ProcessScriptParam and its sister function. We now have a
better understanding of how scripting opcode parameters work,
and how the two special bits they contain affect their value.