Add new fixed-point types (#257)
There's multiple places in the repo that refer to various fixed-point
value formats, but they are usually documented as integers or shorts.
I thought it could be a good idea to document them explicitly, so I
defined new types for them. struct fx64 has also been renamed to
struct fx64_16 for consistency.