Add GitHub Actions check for forgotten header prototypes (#349)
Add a safety check for when a symbol is newly added to the symbol tables
but not to the headers. This is often caused by mistakes rather than
intentional omissions.

Like the symbol compatibility check, this is an optional check that can
be ignored if needed, since sometimes we really may not know what to put
in the headers.